Pr updates made the helper methods all internal
